Upgrade the spec repo to consistent language
Prior to this change, search and hunt were used interchangeably. Since
sail is now the term for looking at water locations and search is now
the term for looking at land locations, hunt is not the proper way to
describe things in the spec. Water and land are used to describe the
location types, so Sail -> Water and Search -> Land. This was cleared up
all through the spec.

Additionally, the spec cleaned up some of the stats labels along the
same line as the sail/search changes, so outputs of solved puzzles had
to be changed.